% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonsaBayesFuns.R
\name{bayes_find_quantile}
\alias{bayes_find_quantile}
\title{bayes_find_quantile}
\usage{
bayes_find_quantile(y, p, sigma, tau, q)
}
\arguments{
\item{y}{value of y}

\item{p}{prior probability of Gaussian slab signal}

\item{sigma}{standard deviation of sampling distribution}

\item{tau}{standard deviation / scale of Gaussian slab}

\item{q}{desired quartile}
}
\value{
quartile from Bayesian posterior distribution
}
\description{
Find quantile of Bayesian posterior distribution for normal
sampling model, and spike-and-slab prior for theta
}
\author{
Spencer Woody
}
