% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtnorm.R
\name{dtnorm}
\alias{dtnorm}
\title{dtnorm}
\usage{
dtnorm(x, mean, sd, t, log.p = FALSE)
}
\arguments{
\item{x}{vector of values at which to evaluate density}

\item{mean}{vector of means}

\item{sd}{vector of standard deviations}

\item{t}{truncation value}

\item{log.p}{logical; if TRUE, probability given as log(p)}
}
\value{
The density of the double-truncated normal distribution at x
}
\description{
Density for the doulbe-truncated normal distribution, x ~ N(mean,
sd^2) * 1(abs(x) > t)
}
\author{
Spencer Woody
}
