% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebayesFuns.R
\name{eb_fun_knownsigma}
\alias{eb_fun_knownsigma}
\title{eb_fun}
\usage{
eb_fun_knownsigma(y, sigma)
}
\arguments{
\item{y}{vector of oberservation}
}
\value{
MMLE estimates of hyperparameters, the hessian matrix at
  the MMLE estimate, and the convergence message
}
\description{
Find empirical Bayes estimates of hyperparameters of spike-slab
prior by using marginal likelihood maximization
}
\details{
Theta follows the spike-slab prior
}
\author{
Spencer Woody
}
