% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efron.R
\name{efron}
\alias{efron}
\title{Efron}
\usage{
efron(z, nmids = 150, pct = -0.01, pct0 = 0.25, df = 10,
  nulltype = "theoretical")
}
\arguments{
\item{z}{}

\item{nmids}{}

\item{pct}{}

\item{pct0}{}

\item{df}{}

\item{nulltype}{}
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Spencer Woody
}
