% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAB.R
\name{FAB}
\alias{FAB}
\title{Create FAB intervals, given the spending function}
\usage{
FAB(theta, w, sigma, alpha = 0.1, yMin = -5, yMax = 5, yNum = 5000,
  verbose = TRUE)
}
\arguments{
\item{theta}{}

\item{w}{}

\item{sigma}{}

\item{alpha}{}

\item{yMin}{}

\item{yMax}{}

\item{yNum}{}

\item{verbose}{}
}
\description{
Create FAB intervals, given the spending function
}
\author{
Spencer Woody
}
