% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_fun_gauss.R
\name{int_fun_gauss}
\alias{int_fun_gauss}
\title{Gotta think of a good description for this...}
\usage{
int_fun_gauss(theta, y, sigma, tau, t)
}
\arguments{
\item{theta}{mean of spike slab}

\item{y}{value of y}

\item{sigma}{standard deviation of sampling distribution}

\item{tau}{standard deviation of slab normal distribution}

\item{t}{truncation point}
}
\description{
Gotta think of a good description for this...
}
