% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABfuns.R
\name{make_w_fab}
\alias{make_w_fab}
\title{make_w_fab}
\usage{
make_w_fab(sigma, p, tau, alpha = 0.1, theta_min = -7, theta_max = 7,
  num_theta = 5000, epsilon = 1e-05, verbose = F)
}
\arguments{
\item{sigma}{standard deviation of sampling distribution}

\item{p}{prior probabilitiy of signal}

\item{tau}{standard deviation of Gaussian slab}

\item{alpha}{confidence level}

\item{theta_min}{minimal value of theta to create spending function}

\item{theta_max}{maximal value of theta to create spending function}

\item{num_theta}{number of thetas over which to create grid}

\item{epsilon}{tolerance level}

\item{verbose}{logical; if TRUE, print progress bar}
}
\value{
a vector of w's, corresponding to the spending function
}
\description{
Make spending function for FAB setting, with spike-slab prior
}
\author{
Spencer Woody
}
