% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_w_theta.R
\name{make_w_theta}
\alias{make_w_theta}
\title{Create optimal spending function for the specified marginal distribution}
\usage{
make_w_theta(marginal_fun, sigma, t, ..., alpha = 0.05, theta_min = -7,
  theta_max = 7, num_theta = 5000, epsilon = 1e-10,
  verbose = FALSE)
}
\arguments{
\item{marginal_fun}{function for marginal density of y}

\item{sigma}{standard deviation for sampling model}

\item{t}{truncation point}

\item{...}{additional arguments to marginal function}

\item{alpha}{confidence level}

\item{theta_min}{minimal theta to create spending function}

\item{theta_max}{maximum theta to create spending function}

\item{num_theta}{length of theta grid}

\item{epsilon}{starting point for root finding}
}
\description{
Create optimal spending function for the specified marginal distribution
}
