% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_funs_joint.R
\name{marginal_gauss_joint}
\alias{marginal_gauss_joint}
\title{marginal_gauss_joint}
\usage{
marginal_gauss_joint(y, sigma, t, tau)
}
\arguments{
\item{y}{value of y}

\item{sigma}{standard deviation}

\item{t}{truncation point}

\item{tau}{scale / standard deviation of Gaussian prior}
}
\value{
Marginal density of y for Gaussian prior on mean theta
}
\description{
Marginal density of y for Gaussian prior on mean theta, under
joint selection
}
\author{
Spencer Woody
}
