% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_funs_conditional.R
\name{marginal_gaussnc_conditional}
\alias{marginal_gaussnc_conditional}
\title{non zero mean Gaussian}
\usage{
marginal_gaussnc_conditional(y, sigma, mu, tau, t)
}
\arguments{
\item{y}{}

\item{sigma}{}

\item{mu}{}

\item{tau}{}

\item{t}{}
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Spencer Woody
}
