% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_funs_conditional.R
\name{marginal_hs_conditional}
\alias{marginal_hs_conditional}
\title{marginal_hs_conditional}
\usage{
marginal_hs_conditional(y, sigma, tau, t)
}
\arguments{
\item{y}{value of y}

\item{sigma}{standard deviation of sampling distribution}

\item{tau}{(global) scale of horse}

\item{t}{truncation point}
}
\value{
marginal density
}
\description{
Marginal density of y for horseshoe prior on mean theta, under
conditional selection
}
\details{
Uses closed-form lower bound density of horseshoe, as described by
Carvalho et al. (2008)
}
\author{
Spencer Woody
}
