% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_funs_joint.R
\name{marginal_hs_joint}
\alias{marginal_hs_joint}
\title{marginal_hs_joint}
\usage{
marginal_hs_joint(y, sigma, tau, t)
}
\arguments{
\item{y}{value of y}

\item{sigma}{standard deviation of sampling distribution}

\item{tau}{(global) scale of horseshoe prior}

\item{t}{truncation point}
}
\value{
marginal density of y for horseshoe prior on mean theta
}
\description{
marginal density of y for horseshoe prior on mean theta, under
joint selection
}
\details{
Uses lower bound of horseshoe density, from Carvahlo et al. (2008)
}
\author{
Spencer Woody
}
