% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_funs_conditional.R
\name{marginal_laplace_conditional}
\alias{marginal_laplace_conditional}
\title{marginal_laplace_conditional}
\usage{
marginal_laplace_conditional(y, sigma, tau, t)
}
\arguments{
\item{y}{value of y}

\item{sigma}{standard deviation of sampling distribution}

\item{tau}{scale of Laplacian prior for theta}

\item{t}{truncation point}
}
\value{
marginal density
}
\description{
Marginal density of y for Laplacian prior on mean theta, under
conditional selection
}
\author{
Spencer Woody
}
