% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_funs_joint.R
\name{marginal_laplace_joint}
\alias{marginal_laplace_joint}
\title{marginal_laplace_joint}
\usage{
marginal_laplace_joint(y, sigma, t, tau)
}
\arguments{
\item{y}{value of y}

\item{sigma}{standard deviation for sampling model}

\item{t}{truncation point}

\item{tau}{scale for laplacian prior of theta}
}
\value{
density of marginal of y for laplacian prior
}
\description{
marginal density of y for Laplacian prior on mean theta, under
joint selection
}
\author{
Spencer Woody
}
