% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_ss_joint1.R
\name{marginal_ss_joint1}
\alias{marginal_ss_joint1}
\title{Density of marginal distribution of y ~ N(theta, sigma^2),
where theta has spike-slab prior, under joint selection}
\usage{
marginal_ss_joint1(y, sigma, t, p, mu, tau)
}
\arguments{
\item{y}{value at which to evaluate}

\item{sigma}{standard deviation of sampling distribution}

\item{t}{truncation point}

\item{p}{proportion for spike part}

\item{tau}{standard deviation of normal}
}
\description{
Density of marginal distribution of y ~ N(theta, sigma^2),
where theta has spike-slab prior, under joint selection
}
