% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phat.R
\name{phat}
\alias{phat}
\title{phat}
\usage{
phat(y, p, sigma, tau)
}
\arguments{
\item{y}{value of y}

\item{p}{prior probability of signal (slab)}

\item{sigma}{standard deviation of sampling distribution of y}

\item{tau}{standard deviation / scale of Gaussian slab}
}
\value{
posterior probability of signal
}
\description{
Posterior probability of being a signal, using spike-slab prior
}
\details{
Mixture of point-mass spike at zero, and Gaussian slab
}
\author{
Spencer Woody
}
