% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_fitter.R
\name{pr_fitter}
\alias{pr_fitter}
\title{Use predictive recursion to obtain a}
\usage{
pr_fitter(y, mu0 = NULL, sig0 = NULL, nulltype = "theoretical")
}
\arguments{
\item{y}{sample}

\item{mu0}{foobar}

\item{sig0}{foobar}

\item{nulltype}{foobar}
}
\value{
a list, where my_fit is the fitted marginal or y
}
\description{
Use predictive recursion to obtain a
}
