% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_fitter_nosplit.R
\name{pr_fitter_nosplit}
\alias{pr_fitter_nosplit}
\title{Use predictive recursion to obtain a semiparametric estimate of
marginal of y (multiple observations per subject, without splitting),}
\usage{
pr_fitter_nosplit(y, mu0 = NULL, sig0 = NULL,
  nulltype = "theoretical")
}
\arguments{
\item{y}{sample}

\item{mu0}{foobar}

\item{sig0}{foobar}

\item{nulltype}{foobar}
}
\value{
a list, where my_fit is the fitted marginal or y
}
\description{
Use predictive recursion to obtain a semiparametric estimate of
marginal of y (multiple observations per subject, without splitting),
}
