% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pr_S_cnd.R
\name{Pr_S_cnd}
\alias{Pr_S_cnd}
\title{Probability of selection event, conditonal on theta}
\usage{
Pr_S_cnd(theta, sigma, t)
}
\arguments{
\item{theta}{the given value of mean of sampling distribution for y}

\item{sigma}{standard deviation for y}

\item{t}{truncation point}
}
\description{
Probability of selection event, conditonal on theta
}
