% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prfdr.R
\name{prfdr}
\alias{prfdr}
\title{Predictive recursion FDR}
\usage{
prfdr(z, mu0 = 0, sig0 = 1, control = list())
}
\arguments{
\item{z}{}

\item{mu0}{}

\item{sig0}{}

\item{control}{}
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
James G. Scott
}
