% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtnorm.R
\name{qtnorm}
\alias{qtnorm}
\title{Quantile for the double-truncated normal distribution, x ~ N(mean,
sd^2) * 1(abs(x) > t)}
\usage{
qtnorm(p, mean, sd, t, log.p = FALSE)
}
\arguments{
\item{mean}{vector of means}

\item{sd}{vector of standard deviations}

\item{t}{truncation point}

\item{log.p}{logical; if TRUE, probability given as log(p)}

\item{x}{vector of values at which to evaluate density}
}
\value{
The density of the double-truncated normal distribution at x
}
\description{
Quantile for the double-truncated normal distribution, x ~ N(mean,
sd^2) * 1(abs(x) > t)
}
\examples{
qtnorm(3, 0, 1, 1)
qtnorm(0.5, 0, 1, 1)

}
