% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_umau_intervals1.R
\name{sa_umau_intervals1}
\alias{sa_umau_intervals1}
\title{Selection-adjusted UMAU intervals}
\usage{
sa_umau_intervals1(y, thetaRange = c(-15, 15), sd, t, alpha = 0.1)
}
\arguments{
\item{y}{Value for which to make a confidence interval}

\item{thetaRange}{Range of theta values for which to construct
acceptance region}

\item{sd}{standard deviation of normal}

\item{t}{truncation, i.e., S = {i: |y_i| > t}}

\item{alpha}{Confidence level}
}
\description{
Selection-adjusted UMAU intervals (one-sided selection region)
}
\details{
Selection-adjusted UMAU intervals
}
\author{
Spencer Woody
}
