% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saBayesFuns.R
\name{sabayes_find_quantile}
\alias{sabayes_find_quantile}
\title{sabayes_find_quantile}
\usage{
sabayes_find_quantile(y, p, sigma, tau, t, q)
}
\arguments{
\item{y}{value of y}

\item{p}{prior probability of signal}

\item{sigma}{standard deviation of sampling distribution}

\item{tau}{standard deviation / scale of Gaussian slab}

\item{t}{truncation point}

\item{q}{desired quantile}
}
\value{
Quantile of Bayesian posterior distribution
}
\description{
Find quantile of selection-adjusted Bayesian posterior
distribution (i.e. under conditional selection), using normal
sampling distribution and spike-and-slab prior on means thetas
}
\details{
.. content for \details{} ..
}
\author{
Spencer Woody
}
