\name{saFAB-package}
\alias{saFAB-package}
\alias{saFAB}
\docType{package}
\title{
\packageTitle{saFAB}
}
\description{
\packageDescription{saFAB}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{saFAB}
\packageIndices{saFAB}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{saFAB}

Maintainer: \packageMaintainer{saFAB}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
