% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saFAB.R
\name{saFAB}
\alias{saFAB}
\title{Create selection-adjusted FAB intervals, given the spending function}
\usage{
saFAB(theta, w, t, sigma, alpha = 0.1, yMin = -5, yMax = 5,
  yNum = 5000, verbose = TRUE)
}
\arguments{
\item{theta}{vector of thetas from spending function}

\item{w}{vector for spending function}

\item{t}{truncation point}

\item{alpha}{confidence level}

\item{yMin}{minimal value of y for the confidence interval}

\item{yMax}{maxmimal value of y for the confidence interval}

\item{yNum}{length of vector of y's at which to make confidence interval}

\item{verbose}{logical; if TRUE (default), print progress bars}
}
\description{
Create selection-adjusted FAB intervals, given the spending function
}
