% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABfuns.R
\name{theta_u_fab}
\alias{theta_u_fab}
\title{theta_u_fab}
\usage{
theta_u_fab(theta_vec, w_theta, y, sigma, alpha = 0.1)
}
\arguments{
\item{theta_vec}{vector of thetas}

\item{w_theta}{vector of w's for thetas (i.e., spending function)}

\item{y}{single observation}

\item{sigma}{standard deviation of sampling distribution}

\item{alpha}{confidence level}
}
\value{
upper bound of FAB interval
}
\description{
Upper bound of FAB interval
}
\author{
Spencer Woody
}
