% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalLength.R
\name{totalLength}
\alias{totalLength}
\title{totalLength}
\usage{
totalLength(intervals)
}
\arguments{
\item{intervals}{}
}
\value{
Sum of lengths of disjoint intervals
}
\description{
Sum length of disjoint intervals
}
\author{
Spencer Woody
}
