% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtnorm.R, R/qtnorm.R
\name{dtnorm}
\alias{dtnorm}
\title{Density for the doulbe-truncated normal distribution, x ~ N(mean,
sd^2) * 1(abs(x) > t)}
\usage{
dtnorm(p, mean, sd, t, log.p = FALSE)

dtnorm(p, mean, sd, t, log.p = FALSE)
}
\arguments{
\item{mean}{vector of means}

\item{sd}{vector of standard deviations}

\item{t}{truncation point}

\item{log.p}{logical; if TRUE, probability given as log(p)}

\item{x}{vector of values at which to evaluate density}

\item{log}{logical; if TRUE, probability given as log(p)}

\item{x}{vector of values at which to evaluate density}

\item{mean}{vector of means}

\item{sd}{vector of standard deviations}

\item{t}{truncation point}
}
\value{
The density of the double-truncated normal distribution at x

The density of the double-truncated normal distribution at x
}
\description{
Density for the doulbe-truncated normal distribution, x ~ N(mean,
sd^2) * 1(abs(x) > t)

Quantile for the doulbe-truncated normal distribution, x ~ N(mean,
sd^2) * 1(abs(x) > t)
}
\examples{
dtnorm(3, 0, 1, 1)
dtnorm(0.5, 0, 1, 1)

qtnorm(3, 0, 1, 1)
qtnorm(0.5, 0, 1, 1)

}
