% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalLength.R
\name{totalLength}
\alias{totalLength}
\title{Calculate total length of disjoint intervals}
\usage{
totalLength(intervals)
}
\arguments{
\item{intervals}{vector containing start and end points of intervals}
}
\description{
Calculate total length of disjoint intervals
}
